requirements.txt
crewai>=0.22.0
crewai-tools>=0.2.0
openinference-instrumentation-crewai>=0.1.0
braintrust[otel]>=0.0.139
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
python-dotenv>=1.0.0